Move 'Action' to 'state' package
